service: refactore
